user-step-3
workflow -> job -> step

in the step we are adding comment to the pr body. to post the comment we are using github token and pr url as env, these env variables can be used in the bash{run} command. Github token is automaticaly used by the gh command.
"Welcome to the repository!"